sys_mon_de_pac/requirements.txt
asgiref==3.9.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.6
django-restframework==0.0.1
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
pytz==2025.2
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
zope.interface==8.0.1
